add ClientRPC interface
